refactor(lowy): extract Emanote.View.Tailwind module (version contract seam)
v4 CLI flags, directive syntax, and plugin list now live in one module.
A future v5 bump changes one file, not three.
