+++ emanote/emanote.cabal
@@ -208,6 +208,7 @@ library
     Emanote.View.Feed
     Emanote.View.LiveServerFiles
     Emanote.View.TagIndex
+    Emanote.View.Tailwind
     Emanote.View.TaskIndex
     Emanote.View.Template
